MBS-9031: Replace all usages of memcached with redis
Redis can do everything we're using memcached for (and more), so this
simplifies our setup by dropping memcached as a dependency.

I didn't see any point in using Cache::Redis, because we'd have to write a
wrapper around it anyway (it handles serialization on its own, doesn't have
delete/*_multi methods, etc.), and Cache::Redis itself is already a very small
wrapper around the Redis module, so instead of writing a wrapper around a
wrapper, we can just write a wrapper. But if we were to use Cache::Redis, we'd
have to add a flag to MusicBrainz::Server::CacheWrapper to disable
serialization, and add expiration time support to set_multi. However, another
advantage of having a new wrapper is that we can use Redis-specific features
like mget in get_multi and command pipelining in set_multi, which the generic
wrapper can't do without breaking the in-memory cache.

Sessions and cached entities are now stored in the same server. Because
sessions are in there, we can't enable any new memory limit or LRU eviction,
i.e. how memcached operated. So, I've changed ENTITY_CACHE_TTL from its
default of 0 (no expiration on keys) on standalone servers to a new default of
86400 (1 day).