Remove Catalyst::Plugin::AutoRestart
The module is being removed for three reasons:
 1. Starlet already kills workers for us after a configured number of requests
    (adjusted here).
 2. Proc::ProcessTable can't compile against musl libc.
 3. MBS-4071.

Also, with the settings currently used in production, the logs confirm that
workers *always* get killed after 60 requests (the minimum they're required to
handle), because they always exceed a memory size of 350000000 at that point.
The new Starlet settings allow them to perform up to 90 requests.