Process seeded relationships on the server
Doing this on the client was messy, because any seeded target entities
had to be requested asynchronously before we could display the
linked relationships. On the server, we can just load the target entities
from the database. In general, Perl makes it easier for us to process
and validate the query parameters that are given.