Pass linkType, attributes directly to linkPhrase utils
Pass the required information (linkType, attributes) directly to the
linkPhrase functions rather than the whole relationship itself. This
makes typing easier in the new relationship editor, since the linkType
may be null on relationship objects while editing, but we can't refine
that fact (as in refining the Flow type) outside of linkPhrase.js.