Make DBDefs JS external from build
The motivation here is to eventually compile static resources for
production into our Docker images. The main roadblock to that is that we
have to recompile resources whenever a DBDefs value changes, due to the
DBDefs modules being included bundles.

To resolve this, we now output the client DBDefs values in a global
__MB_DBDefs__ constant on the page. (Refer to clientDBDefsScript.js
under root/layout/components/.)