Remove DBDefs dependency from webpack config
In a subsequent commit, we'll be running compile_resources.sh as part of
the production Docker image build steps. There's no access to DBDefs
values in that case, because DBDefs.pm doesn't exist - it's rendered by
consul-template when the container runs.

The Webpack publicPath setting is now set at runtime, as documented at
https://webpack.js.org/guides/public-path/#on-the-fly