Autocomplete2: Item types
Autocomplete2 items, prior to this patch, were of type 'EntityItem'
(e.g. a CoreEntityT) or an autocomplete-specific item type like an
ActionItem. But this made it impossible to attach other item-specific
metadata like level, separator, disabled, etc. on EntityItem items
without assigning those onto the entity and running afoul of their Flow
types. It was also awkward to refine actions (based on the existence of
an 'action' property) or add new item types.

This commit adds dedicated item types which can be refined based on
their 'type' property. For example, an EntityItem will now be wrapped in
an OptionItem object (with type 'option').