Autocomplete2: Recent items list
Adds a recent items list that pops up when an empty autocomplete field
is focused, similar to the jQuery UI one.

A new feature when compared to the old widget is that static item lists
are supported, too. In cases where the list is a tree (like for
instruments), the recent entities are displayed twice, but flattened in
the recent list.