RecordingEditForm: Don't disable the submit button if the form has errors
Disabling the submit button prevents us from showing pending errors on the
form. It also bypasses the native popups browsers show when required fields
are empty.

The relationship editor and external links editor both also have code to
disable the submit button on submission, so we have to ensure that that code is
not invoked if the submission was prevented.

Note that we are still lacking client-side error messages for the artist credit
editor, so this commit does not fully realize proper error display across the
whole form.