ArtistCreditEditor: Add client-side validation
Mirrors the FormHandler validation in `Form::Field::ArtistCredit` onto the
client, so the user can be informed of errors before submission.

This is achieved with some refactoring that stores the ArtistCreditEditor's
state as an actual `CompoundFieldT`. That lets us use existing utilities for
updating and displaying field errors. And errors that come from the server
(FormHandler) can be received and displayed with the same code, since those are
also just `CompoundFieldT`s. (See references to `initialField`.)